+++ tagger/src/components/image.rs
@@ -1,26 +1,28 @@
 use crate::components::Title;
-
-use super::Render;
-use anyhow::Result;
+use anyhow::{anyhow, Result};
+use image::DynamicImage;
 use ratatui::{
+    buffer::Buffer,
     layout::{Constraint, Direction, Layout, Rect},
     style::{Color, Style},
-    widgets::{Block, Borders},
-    Frame,
+    widgets::{Block, Borders, StatefulWidget, Widget},
 };
 use ratatui_image::{picker::Picker, FilterType, Resize, StatefulImage};
 use std::path::PathBuf;
+use std::sync::{Arc, RwLock};
+use std::thread;
 
 pub(crate) struct Images<'a> {
-    picker: Picker,
+    picker: Arc<RwLock<Picker>>,
     paths: &'a [&'a PathBuf; 2],
 }
 
 impl<'a> Images<'a> {
-    pub(crate) fn new(paths: &'a [&'a PathBuf; 2]) -> Result<Self> {
+    pub(crate) fn new(paths: &'a [&'a PathBuf; 2]) -> Self {
         #[cfg(not(target_os = "windows"))]
-        let mut picker =
-            Picker::from_termios().map_err(|_| anyhow::anyhow!("Failed to get the picker"))?;
+        let mut picker = Picker::from_termios()
+            .map_err(|_| anyhow::anyhow!("Failed to get the picker"))
+            .unwrap();
         #[cfg(target_os = "windows")]
         let mut picker = {
             let mut picker = Picker::new((12, 24));
@@ -28,38 +30,49 @@ impl<'a> Images<'a> {
             picker
         };
         picker.guess_protocol();
-        Ok(Self { picker, paths })
+        Self {
+            picker: Arc::new(RwLock::new(picker)),
+            paths,
+        }
     }
 }
 
-impl<'a> Render for Images<'a> {
-    fn render(&mut self, f: &mut Frame<'_>, area: Rect) -> Result<()> {
+impl<'a> Widget for Images<'a> {
+    fn render(self, area: Rect, buf: &mut Buffer) {
         let chunks = Layout::default()
             .direction(Direction::Horizontal)
             .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
             .split(area);
-        for i in 0..2 {
-            Image {
-                picker: &mut self.picker,
-                path: self.paths[i],
+        let jhs = (0..2)
+            .map(|i| {
+                let path = self.paths[i].clone();
+                let picker = self.picker.clone();
+                thread::spawn(move || -> Result<Image> {
+                    let image = image::open(&path)?;
+                    Ok(Image::new(picker, image, path))
+                })
+            })
+            .collect::<Vec<_>>();
+        for (jh, &chunk) in jhs.into_iter().zip(chunks.iter()) {
+            if let Ok(Ok(w)) = jh.join() {
+                w.render(chunk, buf);
             }
-            .render(f, chunks[i])?
         }
-        Ok(())
     }
 }
 
 pub struct Grid<'a> {
-    picker: Picker,
+    picker: Arc<RwLock<Picker>>,
     paths: &'a [PathBuf],
     heighlight: [bool; 9],
 }
 
 impl<'a> Grid<'a> {
-    pub(crate) fn new(paths: &'a [PathBuf], heighlight: [bool; 9]) -> Result<Self> {
+    pub(crate) fn new(paths: &'a [PathBuf], heighlight: [bool; 9]) -> Self {
         #[cfg(not(target_os = "windows"))]
         let mut picker = ratatui_image::picker::Picker::from_termios()
-            .map_err(|_| anyhow::anyhow!("Failed to get the picker"))?;
+            .map_err(|_| anyhow::anyhow!("Failed to get the picker"))
+            .unwrap();
         #[cfg(target_os = "windows")]
         let mut picker = {
             let mut picker = ratatui_image::picker::Picker::new((12, 24));
@@ -67,16 +80,16 @@ impl<'a> Grid<'a> {
             picker
         };
         picker.guess_protocol();
-        Ok(Self {
-            picker,
+        Self {
+            picker: Arc::new(RwLock::new(picker)),
             paths,
             heighlight,
-        })
+        }
     }
 }
 
-impl<'a> Render for Grid<'a> {
-    fn render(&mut self, f: &mut Frame<'_>, area: Rect) -> Result<()> {
+impl<'a> Widget for Grid<'a> {
+    fn render(self, area: Rect, buf: &mut Buffer) {
         let grid = Layout::default()
             .direction(Direction::Vertical)
             .constraints([
@@ -100,48 +113,72 @@ impl<'a> Render for Grid<'a> {
                     .collect::<Vec<_>>()
             })
             .collect::<Vec<_>>();
-        for (i, path) in self.paths.iter().enumerate() {
+        let jhs = self
+            .paths
+            .iter()
+            .map(|path| {
+                let path = path.clone();
+                let picker = self.picker.clone();
+                thread::spawn(move || -> Result<Image> {
+                    let image = image::open(&path)?;
+                    Ok(Image::new(picker, image, path.clone()))
+                })
+            })
+            .collect::<Vec<_>>();
+        for (i, jh) in jhs.into_iter().enumerate() {
             let block = Block::default()
                 .title(format!("{}", i + 1))
                 .borders(Borders::ALL)
-                .border_style(Style::default().fg(if self.heighlight[i] {
-                    Color::Green
-                } else {
-                    Color::White
+                .border_style(Style::default().fg(match self.heighlight[i] {
+                    true => Color::Green,
+                    false => Color::White,
                 }));
             let inner = block.inner(grid[i]);
-            f.render_widget(block, grid[i]);
-            Image::new(&mut self.picker, path)?.render(f, inner)?;
+            block.render(grid[i], buf);
+            if let Ok(Ok(w)) = jh.join() {
+                w.render(inner, buf);
+            }
         }
-        Ok(())
     }
 }
 
-struct Image<'a> {
-    picker: &'a mut Picker,
-    path: &'a PathBuf,
+struct Image {
+    picker: Arc<RwLock<Picker>>,
+    image: DynamicImage,
+    path: PathBuf,
 }
 
-impl<'a> Image<'a> {
-    pub(crate) fn new(picker: &'a mut Picker, path: &'a PathBuf) -> Result<Self> {
-        Ok(Self { picker, path })
+impl Image {
+    pub(crate) fn new(picker: Arc<RwLock<Picker>>, image: DynamicImage, path: PathBuf) -> Self {
+        Self {
+            picker,
+            image,
+            path,
+        }
     }
 }
 
-impl Render for Image<'_> {
-    fn render(&mut self, f: &mut Frame<'_>, area: Rect) -> Result<()> {
-        fn inner(this: &mut Image<'_>, f: &mut Frame<'_>, area: Rect) -> Result<()> {
-            let dyn_img = image::open(this.path.canonicalize()?)?;
-            let mut image_fit_state = this.picker.new_resize_protocol(dyn_img);
-            let image = StatefulImage::new(None).resize(Resize::Fit(Some(FilterType::Gaussian)));
-            f.render_stateful_widget(image, area, &mut image_fit_state);
+impl Widget for Image {
+    fn render(self, area: Rect, buf: &mut Buffer) {
+        fn inner(this: Image, area: Rect, buf: &mut Buffer) -> Result<()> {
+            let mut picker = this
+                .picker
+                .write()
+                .map_err(|_| anyhow!("Race condition of Picker RwLock"))?;
+            let mut image_fit_state = picker.new_resize_protocol(this.image);
+            let image = StatefulImage::new(None).resize(Resize::Fit(Some(FilterType::Nearest)));
+            image.render(area, buf, &mut image_fit_state);
             Ok(())
         }
-        inner(self, f, area).or_else(|e| {
-            Title {
-                title: format!("{}: {}", self.path.display(), e),
-            }
-            .render(f, area)
-        })
+        let path = self.path.clone();
+        inner(self, area, buf)
+            .or_else(|e| -> Result<()> {
+                Title {
+                    title: format!("{}: {}", path.display(), e),
+                }
+                .render(area, buf);
+                Ok(())
+            })
+            .ok();
     }
 }