+++ tagger/src/components/input.rs
@@ -1,27 +1,24 @@
-use anyhow::Result;
 use ratatui::{
+    buffer::Buffer,
     layout::{Constraint, Direction, Layout, Rect},
-    widgets::{Block, Paragraph},
-    Frame,
+    widgets::{Block, Borders, Paragraph, Widget},
 };
-
-use super::Render;
 
 #[derive(Debug)]
 pub(crate) struct NumInput {
     pub num: usize,
 }
 
-impl Render for NumInput {
-    fn render(&mut self, f: &mut Frame<'_>, area: Rect) -> Result<()> {
+impl Widget for NumInput {
+    fn render(self, area: Rect, buf: &mut Buffer) {
         let chunks = Layout::default()
             .direction(Direction::Vertical)
             .constraints([Constraint::Percentage(50), Constraint::Percentage(50)])
             .split(area);
-        let title = Paragraph::new("Page to go:").block(Block::default());
-        f.render_widget(title, chunks[0]);
-        let input = Paragraph::new(format!("{}", self.num)).block(Block::default());
-        f.render_widget(input, chunks[1]);
-        Ok(())
+        let title = Paragraph::new("Page to go:").block(Block::default().borders(Borders::ALL));
+        title.render(chunks[0], buf);
+        let input =
+            Paragraph::new(format!("{}", self.num)).block(Block::default().borders(Borders::ALL));
+        input.render(chunks[1], buf);
     }
 }