+++ tagger/src/components/title.rs
@@ -1,26 +1,23 @@
-use super::Render;
-use anyhow::Result;
 use ratatui::{
+    buffer::Buffer,
     layout::Rect,
     style::{Color, Style},
     text::Text,
-    widgets::{Block, Borders, Paragraph, Wrap},
-    Frame,
+    widgets::{Block, Borders, Paragraph, Widget, Wrap},
 };
 
 pub(crate) struct Title {
     pub title: String,
 }
 
-impl Render for Title {
-    fn render(&mut self, f: &mut Frame<'_>, area: Rect) -> Result<()> {
+impl Widget for Title {
+    fn render(self, area: Rect, buf: &mut Buffer) {
         let title_block = Block::default()
             .borders(Borders::ALL)
             .style(Style::default());
         let title = Paragraph::new(Text::styled(&self.title, Style::default().fg(Color::Green)))
             .block(title_block)
             .wrap(Wrap { trim: true });
-        f.render_widget(title, area);
-        Ok(())
+        title.render(area, buf);
     }
 }